requirements.txt
fastapi
uvicorn
gitpython
pydantic
pytest
python-multipart
